pt: Prefetch instruction buffer
Batch reading the instruction buffer from the underlying
dso cache. Previously each instruction would
read it separately. Now we always get a large
buffer and try to reuse it over calls.

In addition I changed the code to use PT MODE.Exec
information to determine the run-mode instead
of asking perf. That is both more correct and more
efficient. It avoids unnecessary map lookups in
this hot path.

This improves decoding performance on a couple
of traces by 2-3%

Signed-off-by: Andi Kleen <ak@linux.intel.com>