perf: Allow guest PEBS for KVM owned counters v2
Currently perf unconditionally disables PEBS for guest.

An upcoming patch adds PEBS virtualization to KVM.
With that a guest can own a PEBS counter.

Perf needs to know that the counter is owned by the guest,
because it:
- Enable it when entering the guest execution
- Not disable it as it would do today.
- Inject the PMI into the guest as needed (see followon patch)

This patch adds a state bit into the perf counter.
The state bit has to be set during counter setup
to avoid races with the PMI. I implemented this
with a new argument to perf_event_create_kernel_counter.

The bit is only set by KVM and cannot be selected
by anyone else.

Then change the MSR entry/exit list to allow
PEBS for these counters.

v2: Fix warning on 32bit. More description.
Signed-off-by: Andi Kleen <ak@linux.intel.com>