locking, x86, tsx: Make irqs_disabled true inside transactions
When we're in a elided lock the interrupts are not really disabled.
Instead the transaction would always abort on any interrupt.
The interrupt disabling is "elided" in this case.

Some code inside irq disabling explicitely checks for disabled
interrupts. A common case doing this is sched_clock_cpu().
To avoid patching all these cases make irqs_disabled() return
true.

Signed-off-by: Andi Kleen <ak@linux.intel.com>