Convert shmem inode counting to per cpu counters
With AIM7 on tmpfs running on a lock elision kernel shmem's
free inode counter is a source of frequent aborts.

Convert it to a percpu counter, similar to how the free blocks
are already accounted. I changed it to count up instead
of count down. This actually simplifies some code.

Signed-off-by: Andi Kleen <ak@linux.intel.com>