locking, tsx: Disable elision for shrinking brk
When a brk shrinks it inevitably leads to an aborting TLB flush inside
the mm_sem, which aborts. Since these aborts are very expensive
(as judged in a AIM7 fileserver run) avoid elision in this case.

Better would be to move the flushing outside the lock.

Based on a patch from Kedar Joshi.

Signed-off-by: Andi Kleen <ak@linux.intel.com>