vfs: Turn the unlinked inode counter into a percpu counter
The unlinked open inodes super block counter that was added some time ago
unfortunately becomes a very hot cache line in some delete heavy
high IOPs or tmpfs workloads with enough cores. With TSX it also
causes plenty of aborts.

Turn the atomic into a percpu_counter. It's nearly perfectly
suited for a percpu counter because it is only checked
in very slow paths, like remount.

One change from the original is that the per cpu counters
are only 32bit max per logical CPU, but are accumulated
occasionally into a full 64vbit counter. The original was a
long. This shouldn't be a problem I hope.

Original commit:

commit 7ada4db88634429f4da690ad1c4eb73c93085f0c
Author: Miklos Szeredi <mszeredi@suse.cz>
Date:   Mon Nov 21 12:11:32 2011 +0100

    vfs: count unlinked inodes

    Add a new counter to the superblock that keeps track of unlinked but
    not yet deleted inodes.

Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Miklos Szeredi <mszeredi@suse.cz>
Signed-off-by: Andi Kleen <ak@linux.intel.com>