Add early percpu counters
It's difficult to convert counters that are used befored mem_init to
per cpu counters. The problem is that the percpu allocator is not functional
early. This affects for example the vmstat counters, which are used
early (e.g. by memblock freeing)

This patch implements simple support for uninitialized per cpu counters.
We keep an extra "initialized" flag. When the flag is not set
just accumulate any counts in the global counter (this is ok
because early code is single threaded). The user has to ensure
it is zeroed (usually by making it static or global)

Then later when the percpu counter is initialized define a special
variant that keeps any counts accumulated earlier. We cannot do this
unconditionally as the counter may contain random binary garbage.

This adds some minor checks to the fast paths, but should not
be a significant burden as the to be checked flag is on cachelines
that are already accessed.

Cc: tj@kernel.org
Signed-off-by: Andi Kleen <ak@linux.intel.com>