selinux: Make selinux follow_link RCU safe
selinux is the only security module that hooks security_follow_link.
We need to make it RCU walk safe to avoid all systems using
selinux fall out of the RCU path.

The avc_has_perm_noaudit() code is already using RCU, so it is
safe to use. The only problem is the audit code.

This adds a  avc_has_perm_rcu() function that only does the
checking part (avc_has_perm_noaudit), and then also checks if
auditing is required. If auditing is required it returns -ECHILD
and expects the caller to retry in non RCU mode. The non
RCU call can then do the auditing after rechecking.

Signed-off-by: Andi Kleen <ak@linux.intel.com>