vfs: Add touch_atime_nonblock
touch_atime has a fast path that is nonblocking and RCU safe,
and a slow part that can block. This patch adds a touch_atime_nonblock
that only does the non blocking part, and returns -ECHILD
if blocking may be required.  In this case the caller can retry
in non RCU mode.

Used later for RCU symlink walking to set the atime on the
symlink.

Signed-off-by: Andi Kleen <ak@linux.intel.com>