vfs: Implement RCU safe symlinking
We ran into some performance issues with symlink path walking not
supporting RCU.

Even a relatively small percentage of symlink accesses in a workload can cause
significant cache line in a larger system, dragging down the performance
even of RCU accesses.

It's also very unexpected for users that adding a symlink to a path
makes the operation a lot slower.

This is an initial attempt at making the symlink code only fall out of
RCU when absolutely needed. The file system follow_link vector
has to be RCU safe, so we use a new follow_link_rcu (the old
follow_link can be eventually deprecated)

The main principle is to unlazy and retry on every potentially
blocking operation.  If the unlazy succeeds continue on
the current point, otherwise the top level name resolution
code retries the complete lookup.

For the symlink lookup retries have to be done for:
- the security callback
- the atime update
- the cond resched test
- the inode reference (dentry could turn negative)
- and the follow_link_rcu() itself

We rely on the symlink not changing so that is safe
to parse the symlink string without taking a reference
to the dentry (symlinks do never change, they only get recreated new)

This makes the assumption that the file system specific inode
portion where the symlink is stored is also subject to RCU.

Other than that it operates similar to the rest of the RCU
path. Don't take the reference counts unless falling back
to avoid touching shared cache lines.

Signed-off-by: Andi Kleen <ak@linux.intel.com>