shmem: Implement follow_symlink_rcu in tmpfs
For short symlinks it's really easy. it's already safe because
it just returns the string pointer.

For long symlinks we need to add a retry check before blocking.
This is straight forward too, but requires passing a special flag
around that tells the low level function to return if it would block.

Signed-off-by: Andi Kleen <ak@linux.intel.com>