Tree-sweep: Mark a large variety of follow_link RCU safe
These are all very simple follow_link()s, generally just
extracting a string somewhere and returning it as
nd_set_link().  No blocking operation anywhere.

The only slightly tricky part is that we have
to check d_inode for NULL before using because
the dentry could turn negative. However it is safe
to reference the inode once checked because it is
protected by RCU.

So mark them all as RCU safe.

Signed-off-by: Andi Kleen <ak@linux.intel.com>