xfs: Add RCU fast symlink path
We can do RCU symlink lookups as long as no memory
allocation and no IO is needed, and there is no
blocking on the ilock.

For the ilock add a second path that does a trylock.

For the memory allocation the requirement is that
the symlink is zero terminated already.  The XFS on disk
format does not necessarily zero terminate unfortunely.

But we can assume that small symlinks fitting into the
inode's inline area are zero terminated when their length
is less than than the maximum inline area. This works
because the inode fork is always zeroed on allocation.

In principle this would be also possible for symlinks
which are stored in the inode, but then read separately
into memory. This would need more changes elsewhere
to make sure these are always zero terminated
or find a reliable way to check for zero-terminatedness.

So right now RCU symlinks are limited to 31 characters
(length of inline area minus one)

Signed-off-by: Andi Kleen <ak@linux.intel.com>