Add kmap_nonblock
The RCU symlink code needs a non blocking kmap(), but it cannot use
kmap_atomic(). However it has a (slower) fallback path that can be
used if this is not possible.

This patch adds a kmap_nonblock() that is only defined for
!CONFIG_HIGHMEM. For CONFIG_HIGHMEM

Add kmem_nonblock() for generic architecture and for parisc
(the only architecture with a private kmap)

The only difference to the normal kmap() on !HIGHMEM is that
it doesn't call might_sleep.

Signed-off-by: Andi Kleen <ak@linux.intel.com>