vfs: Use nonblocking kmap for follow_link
With real HIGHMEM kmap may block. We cannot handle
this in RCU, and also cannot use an atomic kmap because
we may need to block later, but cannot refetch the
symlink string later if it becomes invalidated on
blocking.

So always fall back to non RCU walking with HIGHMEM.
Without highmem we use the kmap_nonblock() added earlier.

The implication on x86 is that RCU symlink only works
on 64bit. I think that's acceptable

Signed-off-by: Andi Kleen <ak@linux.intel.com>