vfs: Use stable inode for put_link
In RCU mode the dentry can turn negative during the lookup. This would
be fatal for put_link, as it needs access to the inode->i_op->put_link
to free the link state. We can always rely on the inode staying
around because it is RCU protected, but we need to actually
use a stable pointer to it instead of relying on the dentry.

Change all put_link users to always look up the inode
in advance and then pass the inode to put_link to free it.

It is safe to not check if it is zero because follow_link
guarantees to error out if it became negative later.

Signed-off-by: Andi Kleen <ak@linux.intel.com>